requirements.txt
Flask
python-dotenv
influxdb